+++ src/main/java/Model/Nurse.java
@@ -5,24 +5,24 @@ import java.text.ParseException;
 import java.text.SimpleDateFormat;
 
 public class Nurse {
-    private int ID;
+    private long ID;
     private String fName;
     private String lName;
     private java.sql.Date DOB;
 
-    private int room;
+    private long room;
     private String s_room;
     private boolean emptyRoom;
 
-    private int deptID;
+    private long deptID;
     private String s_deptID;
     private boolean emptyDeptID;
 
-    public int getID() {
+    public long getID() {
         return ID;
     }
 
-    public void setID(int ID) {
+    public void setID(long ID) {
         this.ID = ID;
     }
 
@@ -63,7 +63,7 @@ public class Nurse {
 
     }
 
-    public int getRoom() {
+    public long getRoom() {
         return room;
     }
     public void setS_room(String s_room){
@@ -72,14 +72,14 @@ public class Nurse {
             this.emptyRoom = true;
         }
         else{
-            this.room = Integer.parseInt(s_room);
+            this.room = Long.parseLong(s_room);
         }
     }
-    public void setRoom(int room) {
+    public void setRoom(long room) {
         this.room = room;
     }
 
-    public int getDeptID() {
+    public long getDeptID() {
         return deptID;
     }
 
@@ -89,11 +89,11 @@ public class Nurse {
             this.emptyDeptID = true;
         }
         else{
-            this.deptID = Integer.parseInt(s_deptID);
+            this.deptID = Long.parseLong(s_deptID);
         }
     }
 
-    public void setDeptID(int deptID) {
+    public void setDeptID(long deptID) {
         this.deptID = deptID;
     }
