+++ src/main/java/Model/Patient.java
@@ -5,7 +5,7 @@ import java.text.ParseException;
 import java.text.SimpleDateFormat;
 
 public class Patient {
-    private int ID;
+    private long ID;
     private String fName;
     private String lName;
     private String gender;
@@ -17,11 +17,11 @@ public class Patient {
     private String s_phone;
     private boolean emptyPhone;
 
-    public int getID() {
+    public long getID() {
         return ID;
     }
 
-    public void setID(int ID) {
+    public void setID(long ID) {
         this.ID = ID;
     }
 
@@ -90,7 +90,7 @@ public class Patient {
         return phone;
     }
 
-    public void setPhone(int phone) {
+    public void setPhone(long phone) {
         this.phone = phone;
     }
