+++ src/main/java/Model/Pharmacist.java
@@ -5,24 +5,24 @@ import java.text.ParseException;
 import java.text.SimpleDateFormat;
 
 public class Pharmacist {
-    private int ID;
+    private long ID;
     private String fName;
     private String lName;
     private java.sql.Date DOB;
 
-    private int office;
+    private long office;
     private String s_office;
     private boolean emptyOffice;
 
-    private int deptID;
+    private long deptID;
     private String s_deptID;
     private boolean emptyDeptID;
 
-    public int getID() {
+    public long getID() {
         return ID;
     }
 
-    public void setID(int ID) {
+    public void setID(long ID) {
         this.ID = ID;
     }
 
@@ -64,11 +64,11 @@ public class Pharmacist {
     }
 
 
-    public int getOffice() {
+    public long getOffice() {
         return office;
     }
 
-    public void setOffice(int office) {
+    public void setOffice(long office) {
         this.office = office;
     }
 
@@ -78,15 +78,15 @@ public class Pharmacist {
             this.emptyOffice = true;
         }
         else{
-            this.office = Integer.parseInt(s_office);
+            this.office = Long.parseLong(s_office);
         }
     }
 
-    public int getDeptID() {
+    public long getDeptID() {
         return deptID;
     }
 
-    public void setDeptID(int deptID) {
+    public void setDeptID(long deptID) {
         this.deptID = deptID;
     }
 
@@ -96,7 +96,7 @@ public class Pharmacist {
             this.emptyDeptID = true;
         }
         else{
-            this.deptID = Integer.parseInt(s_deptID);
+            this.deptID = Long.parseLong(s_deptID);
         }
     }
